utils: wrong size of data types on win64
On a 64-bit system, sizeof(unsigned long) depends on the data model
implemented by the system. It's 4 bytes on LLP64 (Windows), 8 bytes on
LP64 (Linux, etc.). see
https://learn.microsoft.com/en-us/windows/win32/winprog64/abstract-data-modelscan

Thus the multiplication of an unsigned long can result in a
buffer-overflow if the size_t-value is too big and the division
(bytes/u) will not reach 0

+++ src/flb_utils.c
@@ -714,8 +714,8 @@ int flb_utils_time_split(const char *time, int *sec, long *nsec)
 void flb_utils_bytes_to_human_readable_size(size_t bytes,
                                             char *out_buf, size_t size)
 {
-    unsigned long i;
-    unsigned long u = 1024;
+    size_t i;
+    size_t u = 1024;
     static const char *__units[] = {
         "b", "K", "M", "G",
         "T", "P", "E", "Z", "Y", NULL
@@ -728,7 +728,7 @@ void flb_utils_bytes_to_human_readable_size(size_t bytes,
         u *= 1024;
     }
     if (!i) {
-        snprintf(out_buf, size, "%lu%s", (long unsigned int) bytes, __units[0]);
+        snprintf(out_buf, size, "%zu%s", bytes, __units[0]);
     }
     else {
         float fsize = (float) ((double) bytes / (u / 1024));
